archive most of main-rescan.ts
